% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DRejectionSampler.R
\name{jointnorm}
\alias{jointnorm}
\title{math4800 project
this file implemnts a funciton to sample from a joint  random variable that is defifned on a square}
\usage{
jointnorm(x, y)
}
\arguments{
\item{x}{input for the first marginal normal distribution}

\item{y}{input for the first marginal normal distribution}
}
\value{
join normal pdf
}
\description{
This function returns a value from a normal joint continus random variable note both x and y
must be the same value to be a valid for this 2D sampling function sense we are sampling from a square
}
