% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DRejectionSampler.R
\name{jointRejectionSampler}
\alias{jointRejectionSampler}
\title{math4800 project
this file implemnts a funciton to sample from a joint  random variable that is defifned on a square}
\usage{
jointRejectionSampler(n, jointpdf, lower, upper, C)
}
\arguments{
\item{n}{number of observations. n must be an integer >0}

\item{jointpdf}{a function that is the join  pdf of the distribution.}

\item{lower}{lower and upper limits of the  of the distribution. Must be finite such that \eqn{P(a \le X \le b \cap a \le Y \le b) = 1}.}

\item{upper}{upper and upper limits of the distribution}

\item{C}{a numeric such that \eqn{f(x,y) \le C} for all values of x.}
}
\value{
data.frame of random deviate pairs
}
\description{
This function returns a data frome with n elemtns from a joint continus random variable
}
\examples{
jointRejectionSampler(n = 1000, jointpdf = jointnorm, lower = 0, upper = 1, C = 1)
jointRejectionSampler(n = 1000, jointpdf = jointunif, lower = 0, upper = 1, C = 1)
}
