% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DRejectionSampler.R
\name{jointunif}
\alias{jointunif}
\title{math4800 project
this file implemnts a funciton to sample from a joint  random variable that is defifned on a square}
\usage{
jointunif(x, y)
}
\arguments{
\item{x}{input for the first marginal uniform distribution}

\item{y}{input for the first marginal uniform distribution}
}
\value{
join uniform pdf
}
\description{
This function returns a value from a nuiform joint continus random variable
}
