% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{mypdf}
\alias{mypdf}
\title{math4800 project}
\usage{
mypdf(x)
}
\arguments{
\item{x}{number that should be between 0 and 2 or else the pdf is 0}
}
\description{
This function is a simple pdf returns a number
}
