% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{normpdf}
\alias{normpdf}
\title{math4800 project}
\usage{
normpdf(x)
}
\arguments{
\item{x}{number that should be between 0 and 1 or else the pdf is 0}
}
\description{
This function is a simple  normal pdf returns a number
}
