% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{rejectionSampler}
\alias{rejectionSampler}
\title{math4800 project
this file implemnts a funciton to sample from a random variable that is passed in}
\usage{
rejectionSampler(n, pdf, lower, upper, C)
}
\arguments{
\item{n}{number of observations. n must be an integer >0}

\item{pdf}{a function that is the  pdf of the distribution.}

\item{lower}{lower and upper limits of the  of the distribution. Must be finite such that \eqn{P(a \le X \le b \cap a \le Y \le b) = 1}.}

\item{C}{a numeric such that \eqn{f(x,y) \le C} for all values of x.}

\item{lower}{lower and upper limits of the distribution}
}
\value{
data.frame of random deviate pairs
}
\description{
This function returns a vector with n elemtns from a continus random variable
}
